requirements.txt
langchain
langchain-community
langchain-aws
faiss-cpu
pypdf
pymupdf
bs4
streamlit
python-dotenv
cohere
langchain-cohere
langgraph
